+++ apps/wizard/etl_steps/snapshot.py
@@ -3,19 +3,17 @@
 import os
 import subprocess
 import traceback
-from datetime import datetime as dt
 from pathlib import Path
 from typing import Any, Dict, List, Optional, cast
 
 import streamlit as st
-from typing_extensions import Self
 
 from apps.utils.files import generate_step
 from apps.wizard import utils
-from apps.wizard.etl_steps.utils import COOKIE_SNAPSHOT, MD_SNAPSHOT
+from apps.wizard.etl_steps.forms import SnapshotForm
+from apps.wizard.etl_steps.utils import COOKIE_SNAPSHOT, MD_SNAPSHOT, SCHEMA_ORIGIN
 from etl.docs import examples_to_markdown, faqs_to_markdown, guidelines_to_markdown
-from etl.helpers import read_json_schema
-from etl.paths import BASE_DIR, SCHEMAS_DIR, SNAPSHOTS_DIR
+from etl.paths import BASE_DIR, SNAPSHOTS_DIR
 
 #########################################################
 # CONSTANTS #############################################
@@ -24,10 +22,6 @@ st.set_page_config(
     page_title="Wizard: Snapshot",
     page_icon="🪄",
 )
-# Read schema
-SNAPSHOT_SCHEMA = read_json_schema(path=SCHEMAS_DIR / "snapshot-schema.json")
-# Get properties for origin in schema
-schema_origin = SNAPSHOT_SCHEMA["properties"]["meta"]["properties"]["origin"]["properties"]
 # Lists with fields of special types. By default, fields are text inputs.
 FIELD_TYPES_TEXTAREA = [
     "origin.description_snapshot",
@@ -71,143 +65,6 @@ if "snapshot_file" not in st.session_state:
 #########################################################
 # FUNCTIONS & CLASSES ###################################
 #########################################################
-class SnapshotForm(utils.StepForm):
-    """Interface for snapshot form."""
-
-    step_name: str = "snapshot"
-
-    # config
-    namespace: str
-    snapshot_version: str
-    short_name: str
-    file_extension: str
-    is_private: bool
-    dataset_manual_import: bool
-
-    # origin
-    title: str
-    description: str
-    title_snapshot: str
-    description_snapshot: str
-    origin_version: str
-    date_published: str
-    producer: str
-    citation_full: str
-    attribution: Optional[str]
-    attribution_short: str
-    url_main: str
-    url_download: str
-    date_accessed: str
-
-    # license
-    license_url: str
-    license_name: str
-
-    def __init__(self: Self, **data: str | int) -> None:  # type: ignore[reportInvalidTypeVarUse]
-        """Construct form."""
-        # Change name for certain fields (and remove old ones)
-        data["license_url"] = data["origin.license.url"]
-        data["origin_version"] = data["origin.version_producer"]
-        data["dataset_manual_import"] = data["local_import"]
-
-        # Handle custom namespace
-        if "namespace_custom" in data:
-            data["namespace"] = str(data["namespace_custom"])
-
-        # Handle custom license
-        if "origin.license.name_custom" in data:
-            data["license_name"] = data["origin.license.name_custom"]
-        else:
-            data["license_name"] = data["origin.license.name"]
-
-        # Remove unused fields
-        data = {k: v for k, v in data.items() if k not in ["origin.license.url", "origin.license.name"]}
-        # Remove 'origin.' prefix from keys
-        data = {k.replace("origin.", ""): v for k, v in data.items()}
-
-        # Init object (includes schema validation)
-        super().__init__(**data)
-
-        # Handle custom attribution
-        if not self.errors:
-            if "attribution_custom" in data:
-                self.attribution = str(data["attribution_custom"])
-            else:
-                self.attribution = self.parse_attribution(data)
-
-    def parse_attribution(self: Self, data: Dict[str, str | int]) -> str | None:
-        """Parse the field attribution.
-
-        By default, the field attribution contains the format of the attribution, not the actual attribution. This function
-        renders the actual attribution.
-        """
-        attribution_template = cast(str, data["attribution"])
-        if attribution_template == "{producer} ({year})":
-            return None
-        data_extra = {
-            "year": dt.strptime(str(data["date_published"]), "%Y-%m-%d").year,
-            # "version_producer": data["origin_version"],
-        }
-        attribution = attribution_template.format(**data, **data_extra).replace("  ", " ")
-        return attribution
-
-    def validate(self: "SnapshotForm") -> None:
-        """Check that fields in form are valid.
-
-        - Add error message for each field (to be displayed in the form).
-        - Return True if all fields are valid, False otherwise.
-        """
-        # 1) Validate using schema
-        # This only applies to SnapshotMeta fields
-        self.validate_schema(SNAPSHOT_SCHEMA, ["meta"])
-
-        # 2) Check other fields (non meta)
-        fields_required = ["namespace", "snapshot_version", "short_name", "file_extension"]
-        fields_snake = ["namespace", "short_name"]
-        fields_version = ["snapshot_version"]
-
-        self.check_required(fields_required)
-        self.check_snake(fields_snake)
-        self.check_is_version(fields_version)
-
-        # License
-        if self.license_name == "":
-            self.errors["origin.license.name_custom"] = "Please introduce the name of the custom license!"
-
-        # Attribution
-        if self.attribution == "":
-            self.errors["origin.attribution_custom"] = "Please introduce the name of the custom attribute!"
-
-    @property
-    def metadata(self: Self) -> Dict[str, Any]:  # type: ignore[reportIncompatibleMethodOverride]
-        """Define metadata for easy YAML-export."""
-        license_field = {
-            "name": self.license_name,
-            "url": self.license_url,
-        }
-        meta = {
-            "meta": {
-                "origin": {
-                    "title": self.title,
-                    "description": self.description.replace("\n", "\n      "),
-                    "title_snapshot": self.title_snapshot,
-                    "description_snapshot": self.description_snapshot.replace("\n", "\n      "),
-                    "producer": self.producer,
-                    "citation_full": self.citation_full,
-                    "attribution": self.attribution,
-                    "attribution_short": self.attribution_short,
-                    "version_producer": self.origin_version,
-                    "url_main": self.url_main,
-                    "url_download": self.url_download,
-                    "date_published": self.date_published,
-                    "date_accessed": self.date_accessed,
-                    "license": license_field,
-                },
-                "is_public": not self.is_private,
-            }
-        }
-        meta = cast(Dict[str, Any], utils.clean_empty_dict(meta))
-        return meta
 
 
 def _color_req_level(req_level: str) -> str:
@@ -467,7 +324,7 @@ def render_license_field(form: List[Any]) -> List[str]:
     property_name = "origin.license"
     name = "name"
     prop_uri = f"{property_name}.{name}"
-    props_license = schema_origin["license"]
+    props_license = SCHEMA_ORIGIN["license"]
     props = props_license["properties"][name]
     display_name = create_display_name_snap_section(props, name, property_name)
 
@@ -549,7 +406,7 @@ def render_attribution_field(form: List[Any]) -> List[str]:
     parent = "origin"
     name = "attribution"
     prop_uri = f"{parent}.{name}"
-    props = schema_origin[name]
+    props = SCHEMA_ORIGIN[name]
     display_name = create_display_name_snap_section(props, name, parent, title="Attribution format")
 
     # Main decription
@@ -736,17 +593,17 @@ if st.session_state["show_form"]:
         #     "Fill the following fields to help us fill all the created files for you! Note that sometimes some fields might not be available (even if they are labelled as required)."
         # )
         # Get categories
-        for k, v in schema_origin.items():
+        for k, v in SCHEMA_ORIGIN.items():
             if "category" not in v:
                 print(k)
                 print(v)
                 # raise ValueError(f"Category not found for {k}")
-        categories_in_schema = {v["category"] for k, v in schema_origin.items()}
+        categories_in_schema = {v["category"] for k, v in SCHEMA_ORIGIN.items()}
         assert categories_in_schema == set(
             ACCEPTED_CATEGORIES
         ), f"Unknown categories in schema: {categories_in_schema - set(ACCEPTED_CATEGORIES)}"
 
-        form_metadata = render_fields_from_schema(schema_origin, "origin", [], categories=ACCEPTED_CATEGORIES)
+        form_metadata = render_fields_from_schema(SCHEMA_ORIGIN, "origin", [], categories=ACCEPTED_CATEGORIES)
 
         # 3) Submit
         submitted = st.form_submit_button("Submit", type="primary", use_container_width=True, on_click=update_state)
@@ -768,7 +625,7 @@ else:
 #########################################################
 if submitted:
     # Create form
-    form = cast(SnapshotForm, SnapshotForm.from_state())
+    form = SnapshotForm.from_state()
 
     if not form.errors:
         form_widget.empty()
