+++ apps/wizard/etl_steps_old/garden.py
@@ -13,6 +13,7 @@ import etl.grapher_model as gm
 from apps.utils.files import add_to_dag, generate_step_to_channel
 from apps.wizard import utils
 from apps.wizard.etl_steps_old.utils import COOKIE_GARDEN, MD_GARDEN, TAGS_DEFAULT
+from apps.wizard.utils.components import config_style_html
 from etl.config import DB_HOST, DB_NAME
 from etl.db import get_session
 from etl.files import ruamel_dump, ruamel_load
@@ -35,7 +36,7 @@ CURRENT_DIR = Path(__file__).parent
 st.session_state["step_name"] = "garden"
 APP_STATE = utils.AppState()
 # Config style
-utils.config_style_html()
+config_style_html()
 # DUMMY defaults
 dummy_values = {
     "namespace": "dummy",
