+++ apps/wizard/etl_steps_old/grapher.py
@@ -9,6 +9,7 @@ from typing_extensions import Self
 from apps.utils.files import add_to_dag, generate_step_to_channel
 from apps.wizard import utils
 from apps.wizard.etl_steps_old.utils import COOKIE_GRAPHER, MD_GRAPHER
+from apps.wizard.utils.components import config_style_html
 from etl.paths import DAG_DIR
 
 #########################################################
@@ -28,7 +29,7 @@ CURRENT_DIR = Path(__file__).parent
 st.session_state["step_name"] = "grapher"
 APP_STATE = utils.AppState()
 # Config style
-utils.config_style_html()
+config_style_html()
 # DUMMY defaults
 dummy_values = {
     "namespace": "dummy",
