+++ apps/wizard/etl_steps_old/meadow.py
@@ -11,6 +11,7 @@ from typing_extensions import Self
 from apps.utils.files import add_to_dag, generate_step_to_channel
 from apps.wizard import utils
 from apps.wizard.etl_steps_old.utils import COOKIE_MEADOW, MD_MEADOW, load_datasets
+from apps.wizard.utils.components import config_style_html
 from etl.paths import BASE_DIR, DAG_DIR
 from etl.steps import load_from_uri
 
@@ -30,7 +31,7 @@ CURRENT_DIR = Path(__file__).parent
 st.session_state["step_name"] = "meadow"
 APP_STATE = utils.AppState()
 # Config style
-utils.config_style_html()
+config_style_html()
 # DUMMY defaults
 dummy_values = {
     "namespace": "dummy",
