+++ apps/wizard/utils/defaults.py
@@ -46,7 +46,7 @@ def create_wizard_defaults(defaults: Dict[str, Any], overwrite: bool = False) ->
             json.dump(defaults, f)
 
 
-def update_wizard_defaults_from_form(form: StepForm) -> None:
+def update_wizard_defaults_from_form(form: Any) -> None:
     """Update wizard defaults file."""
     # Load config
     config = load_wizard_defaults()
